fixed checkstyle, added tests

+++ src/test/resources/database/delete/delete-data.sql
@@ -1,4 +1,4 @@
 DELETE FROM rentals;
 DELETE FROM cars;
-DELETE FROM users;
-DELETE FROM user_role;
+DELETE FROM users where id != 1;
+DELETE FROM user_role where user_id != 1;
